+++ frontend/components/SenatorPortrait.tsx
@@ -6,7 +6,7 @@ import Faction from "@/classes/Faction"
 import styles from "./SenatorPortrait.module.css"
 import Title from "@/classes/Title"
 import TitleIcon from "@/components/TitleIcon"
-import SelectedDetail from "@/types/SelectedDetail"
+import SelectedDetail from "@/types/SelectedDetail.temp"
 import factionColors from "@/data/factionColors.json"
 import FactionLeaderPattern from "@/images/patterns/factionLeader.svg"
 import DeadIcon from "@/images/icons/dead.svg"
