+++ frontend/contexts/GameContext.tsx
@@ -13,7 +13,7 @@ import Collection from "@/classes/Collection"
 import Title from "@/classes/Title"
 import Game from "@/classes/Game"
 import Step from "@/classes/Step"
-import SelectedDetail from "@/types/SelectedDetail"
+import SelectedDetail from "@/types/SelectedDetail.temp"
 import ActionLog from "@/classes/ActionLog"
 import SenatorActionLog from "@/classes/SenatorActionLog"
 import Phase from "@/classes/Phase"
