+++ frontend/types/SelectedDetail.temp.ts
